requirements.txt
faiss-cpu
sentence-transformers
pandas
memory-profiler
line-profiler
path
python-multipart
dill
fastparquet
from-root
pip-chill
path
PyYAML
passlib
pymongo
requests
pillow
dnspython
tqdm
openpyxl
boto3
-e.